+++ spec/e2e/pbivizNewSpec.js
@@ -29,8 +29,6 @@
 const fs = require('fs-extra');
 const path = require('path');
 const utils = require('./utils');
-const lodashDifference = require('lodash.difference');
-const lodashFindIndex = require('lodash.findindex');
 
 let FileSystem = require('../helpers/FileSystem.js');
 const writeMetadata = require("./utils").writeMetadata;
@@ -74,7 +72,7 @@ describe("E2E - pbiviz new", () => {
         expectedFiles.concat(utils.readdirSyncRecursive(path.join(templatePath, 'visuals', '_global')));
         expectedFiles.push('/pbiviz.json');
         let visualFiles = utils.readdirSyncRecursive(visualPath);
-        let fileDiff = lodashDifference(expectedFiles, visualFiles);
+        let fileDiff = [expectedFiles, visualFiles].reduce((a, b) => a.filter(c => !b.includes(c)));
         expect(fileDiff.length).toBe(0);
 
         // check exists node_modules directory
@@ -392,7 +390,7 @@ describe("E2E - pbiviz new", () => {
 
             //tsconfig should've been updated
             let tsConfig = fs.readJsonSync(path.join(visualPath, 'tsconfig.json'));
-            let typeDefIndex = lodashFindIndex(tsConfig.files, i => i.match(/.api\/.+\/PowerBI-visuals.d.ts$/));
+            let typeDefIndex = tsConfig.files.findIndex(i => i.match(/.api\/.+\/PowerBI-visuals.d.ts$/));
             expect(tsConfig.files[typeDefIndex]).toBe('.api/v1.0.0/PowerBI-visuals.d.ts');
 
             //.vscode/settings.json should've been set to the correct schemas